requirements.txt
Django==1.11.1
Pillow
django-crispy-forms
django-registration-redux==1.6

